generalize ShaderModel

+++ src/include/miopen/solver.hpp
@@ -2450,7 +2450,7 @@ struct ConvAsmBwdWrW3x3 final : ConvTunableSolver<PerformanceConfigAsmDirect3x3W
                              const PerformanceConfigAsmDirect3x3WrW& config) const override;
 };
 
-template <size_t Winodata, size_t Winofilter>
+template <uint32_t Winodata, uint32_t Winofilter>
 struct ConvWinoFuryRxS final : ConvSolver
 {
     const std::string& SolverDbId() const override
@@ -2462,8 +2462,7 @@ struct ConvWinoFuryRxS final : ConvSolver
     bool IsDynamic() const override { return true; }
     float GetWti(const ConvolutionContext&, const ProblemDescription&) const override;
 
-    ConvSolution GetSolution(const ConvolutionContext&,
-                             const ProblemDescription&) const override;
+    ConvSolution GetSolution(const ConvolutionContext&, const ProblemDescription&) const override;
 
     static constexpr bool is2x3() { return Winodata == 2 && Winofilter == 3; }
     static constexpr bool is3x2() { return Winodata == 3 && Winofilter == 2; }
